refactor: extract testbench and runtime crates
