+++ crates/celox-runtime/src/event_buffer.rs
