+++ crates/celox-testbench/Cargo.toml
@@ -9,6 +9,7 @@ edition.workspace     = true
 
 [dependencies]
 num-bigint = { workspace = true }
+num-traits = { workspace = true }
 
 [lints]
 workspace = true
