--- crates/celox/src/context_width.rs
@@ -1,1 +1,0 @@
-pub use celox_frontend_veryl::context_width::*;
