+++ crates/celox/src/testbench.rs
@@ -5,26 +5,21 @@
 //! memory buffer.  Signals ≤64 bits use native `u64` arithmetic with zero
 //! heap allocation; wider signals fall back to `BigUint`.
 
-use crate::backend::get_byte_size;
 use crate::backend::memory_layout::{
     RUNTIME_EVENT_HEADER_SIZE, RUNTIME_EVENT_SLOT_ARG_COUNT_OFFSET,
     RUNTIME_EVENT_SLOT_PAYLOAD_OFFSET, RUNTIME_EVENT_SLOT_SEQ_OFFSET,
     RUNTIME_EVENT_SLOT_SITE_OFFSET, RUNTIME_EVENT_WRITING,
 };
 use crate::backend::traits::SimBackend;
-use crate::context_width::{
-    ValueContext, binary_semantics, cast_semantics, expression_signed, get_expr_width,
-};
-use crate::display_format::{DisplayFormatArg, format_display_arg};
-use crate::ir::{AbsoluteAddr, Program, RuntimeEventKind, RuntimeEventSite, SignalRef};
+use crate::ir::{AbsoluteAddr, Program, SignalRef};
 use crate::simulator::{RuntimeEvent, RuntimeFormatContext, Simulator};
 use celox_frontend_veryl::VerylTestbenchSource;
 pub use celox_testbench::SourceLocation;
 use celox_testbench::{
-    AssertMessage as GenericAssertMessage, ClockCount as GenericClockCount, ExprBytecode,
-    ExprOpcode as TbOpcode, LoopBound as GenericLoopBound, SemanticArgument, SemanticSignal,
-    SemanticStatement, StateLocation, TestbenchOperator as Op, TestbenchProgram,
-    TestbenchStatement as GenericTestbenchStatement,
+    DisplayFormatArg, ExecutableArgument, ExecutableAssertMessage, ExecutableClockCount,
+    ExecutableLoopBound, ExecutableStatement, ExecutableTestbench, TestbenchOperator as Op,
+    TestbenchProgram, TestbenchStatement as GenericTestbenchStatement, TestbenchValue as TbValue,
+    format_display_arg,
 };
 use num_bigint::{BigInt, BigUint, Sign};
 use num_traits::ToPrimitive as _;
@@ -32,15 +27,6 @@ use std::sync::{
     OnceLock,
     atomic::{AtomicU64, Ordering},
 };
-use veryl_analyzer::ir::{
-    ArrayLiteralItem, AssertKind, CasePattern, Expression, Factor, ForBound, ForRange, Function,
-    FunctionCall, Op as VerylOp, Statement, SystemFunctionInput, SystemFunctionKind, TbMethod,
-    TbMethodCall, VarId,
-};
-use veryl_analyzer::value::byte_value_to_string;
-use veryl_parser::resource_table::{self, StrId};
-
-type UnboundTbOpcode = TbOpcode<StateLocation<AbsoluteAddr>>;
 
 // ── Public types ───────────────────────────────────────────────────────
 
@@ -67,9 +53,7 @@ pub struct TestResultDetailed {
 }
 
 /// Opaque, precompiled native testbench program for a built simulator.
-pub struct CompiledTestbench<B: SimBackend> {
-    pub(crate) stmts: Vec<TestbenchStatement<B>>,
-}
+pub type CompiledTestbench<B> = ExecutableTestbench<<B as SimBackend>::Event, SignalRef>;
 
 /// Result of executing a testbench with an optional tick limit.
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -79,759 +63,29 @@ pub struct LimitedTestbenchResult {
     pub tick_limit_reached: bool,
 }
 
-pub(crate) type AssertMessage = GenericAssertMessage<CompiledAssertArg>;
+pub(crate) type AssertMessage = ExecutableAssertMessage;
 
 /// Clock cycle count: either a compile-time constant or a runtime expression.
-pub type ClockCount = GenericClockCount<CompiledExpr>;
+pub type ClockCount = ExecutableClockCount;
 
-pub type LoopBound = GenericLoopBound<CompiledExpr>;
+pub type LoopBound = ExecutableLoopBound;
 
-pub(crate) type TestbenchStatement<B> =
-    GenericTestbenchStatement<<B as SimBackend>::Event, SignalRef, CompiledExpr, CompiledAssertArg>;
+pub(crate) type TestbenchStatement<B> = ExecutableStatement<<B as SimBackend>::Event, SignalRef>;
 
-pub(crate) struct CompiledAssertArg {
-    expr: CompiledExpr,
-    width: usize,
-    signed: bool,
-    is_string: bool,
-}
-
-// ── Bytecode VM ────────────────────────────────────────────────────────
-
-/// A compiled expression: flat bytecode evaluated on a stack VM.
-#[derive(Debug)]
-pub struct CompiledExpr {
-    bytecode: ExprBytecode,
-}
-
-/// Stack value: either a native u64 or a heap-allocated BigUint.
-#[derive(Clone, Debug)]
-enum TbValue {
-    U64(u64),
-    Wide(BigUint),
-}
-
-impl TbValue {
-    #[inline]
-    fn to_u64(&self) -> u64 {
-        match self {
-            TbValue::U64(v) => *v,
-            TbValue::Wide(v) => {
-                let digits = v.to_u64_digits();
-                digits.first().copied().unwrap_or(0)
-            }
-        }
-    }
-
-    #[inline]
-    fn is_zero(&self) -> bool {
-        match self {
-            TbValue::U64(v) => *v == 0,
-            TbValue::Wide(v) => *v == BigUint::ZERO,
-        }
-    }
-
-    #[inline]
-    fn to_biguint(&self) -> BigUint {
-        match self {
-            TbValue::U64(v) => BigUint::from(*v),
-            TbValue::Wide(v) => v.clone(),
-        }
-    }
-}
-
-impl CompiledExpr {
-    /// Evaluate against raw simulator memory, returning the result as u64.
-    /// For wide results, returns the low 64 bits.
-    pub fn eval_u64(&self, memory: *mut u8) -> u64 {
-        self.eval(memory).to_u64()
-    }
-
-    /// Evaluate and return the full `TbValue` (preserves wide results).
-    fn eval_value(&self, memory: *mut u8) -> TbValue {
-        self.eval(memory)
-    }
-
-    pub fn eval_bool(&self, memory: *mut u8) -> bool {
-        !self.eval(memory).is_zero()
-    }
-
-    /// Core evaluation loop.  Uses `TbValue` to handle both u64 and wide
-    /// signals on a single stack.  The common case (all ≤64-bit operands)
-    /// stays in the `TbValue::U64` variant and never allocates.
-    fn eval(&self, memory: *mut u8) -> TbValue {
-        let mut stack: Vec<TbValue> = Vec::with_capacity(16);
-        let mut pc: usize = 0;
-        let ops = self.bytecode.ops();
-
-        while pc < ops.len() {
-            self.exec_at(ops, &mut pc, &mut stack, memory);
-        }
-        stack.pop().unwrap_or_else(|| {
-            debug_assert!(false, "testbench bytecode: stack empty after evaluation");
-            TbValue::U64(0)
-        })
-    }
-
-    /// Execute the opcode at `pc` and advance `pc` past it.
-    /// Handles all opcodes including `Ternary` (with recursive sub-block
-    /// evaluation), so there is no separate `step()` function.
-    fn exec_at(&self, ops: &[TbOpcode], pc: &mut usize, stack: &mut Vec<TbValue>, memory: *mut u8) {
-        match &ops[*pc] {
-            TbOpcode::ConstU64(v) => {
-                stack.push(TbValue::U64(*v));
-                *pc += 1;
-            }
-            TbOpcode::ConstWide(v) => {
-                stack.push(TbValue::Wide(v.clone()));
-                *pc += 1;
-            }
-            TbOpcode::LoadU64 {
-                location,
-                byte_size,
-                mask,
-            } => {
-                // SAFETY: caller guarantees `memory` is valid simulator memory
-                let val = unsafe { read_le_u64(memory.add(*location), *byte_size) } & mask;
-                stack.push(TbValue::U64(val));
-                *pc += 1;
-            }
-            TbOpcode::LoadWide {
-                location,
-                byte_size,
-                width,
-            } => {
-                let val = unsafe { read_le_wide(memory.add(*location), *byte_size, *width) };
-                stack.push(TbValue::Wide(val));
-                *pc += 1;
-            }
-            TbOpcode::BinOp(op) => {
-                let r = stack.pop().unwrap_or_else(|| {
-                    debug_assert!(false, "testbench bytecode: BinOp rhs underflow");
-                    TbValue::U64(0)
-                });
-                let l = stack.pop().unwrap_or_else(|| {
-                    debug_assert!(false, "testbench bytecode: BinOp lhs underflow");
-                    TbValue::U64(0)
-                });
-                stack.push(eval_binop(l, *op, r));
-                *pc += 1;
-            }
-            TbOpcode::TypedBinOp {
-                op,
-                lhs_width,
-                rhs_width,
-                result_width,
-                lhs_signed,
-                rhs_signed,
-            } => {
-                let r = stack.pop().unwrap_or_else(|| {
-                    debug_assert!(false, "testbench bytecode: TypedBinOp rhs underflow");
-                    TbValue::U64(0)
-                });
-                let l = stack.pop().unwrap_or_else(|| {
-                    debug_assert!(false, "testbench bytecode: TypedBinOp lhs underflow");
-                    TbValue::U64(0)
-                });
-                stack.push(eval_typed_binop(
-                    l,
-                    *op,
-                    r,
-                    *lhs_width,
-                    *rhs_width,
-                    *result_width,
-                    *lhs_signed,
-                    *rhs_signed,
-                ));
-                *pc += 1;
-            }
-            TbOpcode::TypedUnary {
-                op,
-                operand_width,
-                result_width,
-            } => {
-                if let Some(top) = stack.last_mut() {
-                    *top = eval_typed_unop(*op, top, *operand_width, *result_width);
-                } else {
-                    debug_assert!(false, "testbench bytecode: TypedUnary underflow");
-                }
-                *pc += 1;
-            }
-            TbOpcode::Resize {
-                source_width,
-                target_width,
-                signed,
-            } => {
-                if let Some(top) = stack.last_mut() {
-                    *top = resize_tb_value(top, *source_width, *target_width, *signed);
-                } else {
-                    debug_assert!(false, "testbench bytecode: Resize underflow");
-                }
-                *pc += 1;
-            }
-            TbOpcode::ConcatPart {
-                part_width,
-                result_width,
-            } => {
-                let part = stack.pop().unwrap_or_else(|| {
-                    debug_assert!(false, "testbench bytecode: ConcatPart value underflow");
-                    TbValue::U64(0)
-                });
-                let accumulator = stack.pop().unwrap_or_else(|| {
-                    debug_assert!(
-                        false,
-                        "testbench bytecode: ConcatPart accumulator underflow"
-                    );
-                    TbValue::U64(0)
-                });
-                if let (TbValue::U64(accumulator), TbValue::U64(part)) = (&accumulator, &part)
-                    && *result_width <= 64
-                {
-                    let shifted = if *part_width >= 64 {
-                        0
-                    } else {
-                        accumulator << part_width
-                    };
-                    stack.push(TbValue::U64(shifted | (part & width_mask_u64(*part_width))));
-                } else {
-                    let value = (accumulator.to_biguint() << part_width)
-                        | normalized_bits(&part, *part_width);
-                    stack.push(tb_value_from_bits(value, *result_width));
-                }
-                *pc += 1;
-            }
-            TbOpcode::Ternary { then_len, else_len } => {
-                let cond = stack.pop().unwrap_or_else(|| {
-                    debug_assert!(false, "testbench bytecode: Ternary cond underflow");
-                    TbValue::U64(0)
-                });
-                *pc += 1; // skip past Ternary opcode
-                if !cond.is_zero() {
-                    let then_end = *pc + then_len;
-                    while *pc < then_end {
-                        self.exec_at(ops, pc, stack, memory);
-                    }
-                    *pc += else_len; // skip else block
-                } else {
-                    *pc += then_len; // skip then block
-                    let else_end = *pc + else_len;
-                    while *pc < else_end {
-                        self.exec_at(ops, pc, stack, memory);
-                    }
-                }
-            }
-            TbOpcode::LoadIndexed {
-                location,
-                stride_bytes,
-                element_byte_size,
-                element_width,
-            } => {
-                let idx = stack.pop().unwrap_or_else(|| {
-                    debug_assert!(false, "testbench bytecode: LoadIndexed underflow");
-                    TbValue::U64(0)
-                });
-                let i = idx.to_u64() as usize;
-                let offset = location + i * stride_bytes;
-                let mask = if *element_width >= 64 {
-                    u64::MAX
-                } else {
-                    (1u64 << element_width) - 1
-                };
-                let val = unsafe { read_le_u64(memory.add(offset), *element_byte_size) } & mask;
-                stack.push(TbValue::U64(val));
-                *pc += 1;
-            }
-            TbOpcode::LoadBitSelect {
-                location,
-                base_byte_size,
-                select_width,
-            } => {
-                let bit_idx = stack.pop().unwrap_or_else(|| {
-                    debug_assert!(false, "testbench bytecode: LoadBitSelect underflow");
-                    TbValue::U64(0)
-                });
-                let shift = bit_idx.to_u64() as usize;
-                let full_val = unsafe { read_le_u64(memory.add(*location), *base_byte_size) };
-                let mask = if *select_width >= 64 {
-                    u64::MAX
-                } else {
-                    (1u64 << select_width) - 1
-                };
-                let val = (full_val >> shift) & mask;
-                stack.push(TbValue::U64(val));
-                *pc += 1;
-            }
-            TbOpcode::StoreU64 {
-                location,
-                byte_size,
-            } => {
-                let val = stack.pop().unwrap_or_else(|| {
-                    debug_assert!(false, "testbench bytecode: StoreU64 underflow");
-                    TbValue::U64(0)
-                });
-                let v = val.to_u64();
-                let bytes = v.to_le_bytes();
-                let n = (*byte_size).min(8);
-                unsafe {
-                    std::ptr::copy_nonoverlapping(bytes.as_ptr(), memory.add(*location), n);
-                }
-                *pc += 1;
-            }
-        }
-    }
-}
-
-fn static_string_expr(expr: &Expression) -> Option<String> {
-    if !expr.comptime().r#type.is_string() {
-        return None;
-    }
-    let value = expr.comptime().get_value().ok()?;
-    byte_value_to_string(value)
-}
-
-fn compile_assert_arg(
-    input: &SystemFunctionInput,
-    ec: &ExprCompiler<'_>,
-) -> SemanticArgument<AbsoluteAddr> {
-    let expr = &input.0;
-    let width = {
-        let ctx_width = expr.comptime().expr_context.width;
-        if ctx_width > 0 {
-            ctx_width
-        } else if let Some(type_width) = expr.comptime().r#type.total_width() {
-            type_width
-        } else if let Ok(value) = expr.comptime().get_value() {
-            value.width()
-        } else {
-            0
-        }
-    };
-    SemanticArgument {
-        expr: ec.compile(expr),
-        width,
-        signed: expression_signed(expr),
-        is_string: expr.comptime().r#type.is_string(),
-    }
-}
-
-fn assert_arg_width(input: &SystemFunctionInput) -> usize {
-    let expr = &input.0;
-    let ctx_width = expr.comptime().expr_context.width;
-    if ctx_width > 0 {
-        ctx_width
-    } else if let Some(type_width) = expr.comptime().r#type.total_width() {
-        type_width
-    } else if let Ok(value) = expr.comptime().get_value() {
-        value.width()
-    } else {
-        0
-    }
-}
-
-fn runtime_event_site_for_assert(
-    kind: &AssertKind,
-    args: &[SystemFunctionInput],
-) -> RuntimeEventSite {
-    let (template, value_args) = if args
-        .first()
-        .and_then(|arg| static_string_expr(&arg.0))
-        .is_some()
-    {
-        (
-            args.first().and_then(|arg| static_string_expr(&arg.0)),
-            &args[1..],
-        )
-    } else {
-        (None, args)
-    };
-    RuntimeEventSite {
-        kind: match kind {
-            AssertKind::Fatal => RuntimeEventKind::AssertFatal,
-            AssertKind::Continue => RuntimeEventKind::AssertContinue,
-        },
-        template,
-        arg_widths: value_args.iter().map(assert_arg_width).collect(),
-        arg_signed: value_args
-            .iter()
-            .map(|arg| expression_signed(&arg.0))
-            .collect(),
-        arg_is_string: value_args
-            .iter()
-            .map(|arg| arg.0.comptime().r#type.is_string())
-            .collect(),
-    }
-}
-
-fn function_body(
-    funcs: &fxhash::FxHashMap<VarId, Function>,
-    fc: &veryl_analyzer::ir::FunctionCall,
-) -> Option<veryl_analyzer::ir::FunctionBody> {
-    let func = funcs.get(&fc.id)?;
-    if let Some(idx) = &fc.index {
-        func.get_function(idx)
-    } else {
-        func.get_function(&[])
-    }
-}
-
-fn collect_runtime_event_sites(
-    stmts: &[Statement],
-    funcs: &fxhash::FxHashMap<VarId, Function>,
-    out: &mut Vec<RuntimeEventSite>,
-) {
-    for stmt in stmts {
-        match stmt {
-            Statement::SystemFunctionCall(sf) => {
-                if let SystemFunctionKind::Assert { kind, args, .. } = &sf.kind {
-                    out.push(runtime_event_site_for_assert(kind, args));
-                }
-            }
-            Statement::If(s) => {
-                collect_runtime_event_sites(&s.true_side, funcs, out);
-                collect_runtime_event_sites(&s.false_side, funcs, out);
-            }
-            Statement::For(s) => collect_runtime_event_sites(&s.body, funcs, out),
-            Statement::FunctionCall(fc) => {
-                if let Some(body) = function_body(funcs, fc) {
-                    collect_runtime_event_sites(&body.statements, funcs, out);
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
-fn count_assert_statements(
-    stmts: &[Statement],
-    funcs: &fxhash::FxHashMap<VarId, Function>,
-) -> usize {
-    let mut count = 0;
-    for stmt in stmts {
-        match stmt {
-            Statement::SystemFunctionCall(sf) => {
-                if matches!(&sf.kind, SystemFunctionKind::Assert { .. }) {
-                    count += 1;
-                }
-            }
-            Statement::If(s) => {
-                count += count_assert_statements(&s.true_side, funcs);
-                count += count_assert_statements(&s.false_side, funcs);
-            }
-            Statement::For(s) => count += count_assert_statements(&s.body, funcs),
-            Statement::FunctionCall(fc) => {
-                if let Some(body) = function_body(funcs, fc) {
-                    count += count_assert_statements(&body.statements, funcs);
-                }
-            }
-            _ => {}
-        }
-    }
-    count
-}
+pub(crate) type CompiledAssertArg = ExecutableArgument;
 
 pub(crate) fn project_observability(program: &mut Program, source: &VerylTestbenchSource) {
-    let Some(stmts) = source.initial_statements.as_ref() else {
-        return;
-    };
-    let mut sites = Vec::new();
-    collect_runtime_event_sites(stmts, &source.functions, &mut sites);
+    let (sites, read_variables) = celox_frontend_veryl::collect_testbench_observability(source);
     program.runtime_schema.runtime_event_sites.extend(sites);
-    program.runtime_schema.testbench_read_roots = initial_read_addresses(program, source);
-}
-
-/// Return top-module signals read by the native testbench. Testbench bytecode
-/// reads these directly from simulator memory, so they are external roots for
-/// dead-store elimination even though no SIR execution unit loads them.
-fn initial_read_addresses(
-    program: &Program,
-    source: &VerylTestbenchSource,
-) -> crate::HashSet<AbsoluteAddr> {
-    let Some(stmts) = source.initial_statements.as_ref() else {
-        return crate::HashSet::default();
-    };
-    let Some(&root_instance_id) = program
-        .frontend
-        .instance_ids
-        .get(&crate::ir::InstancePath(vec![]))
-    else {
-        return crate::HashSet::default();
-    };
-    let Some(&root_module_id) = program.frontend.instance_module.get(&root_instance_id) else {
-        return crate::HashSet::default();
-    };
-    let Some(root_variables) = program.frontend.module_variables.get(&root_module_id) else {
-        return crate::HashSet::default();
-    };
-
-    let mut reads = crate::HashSet::default();
-    let mut active_functions = crate::HashSet::default();
-    collect_statement_reads(stmts, &source.functions, &mut active_functions, &mut reads);
-    reads
+    program.runtime_schema.testbench_read_roots = read_variables
         .into_iter()
-        .filter(|id| root_variables.contains_key(id))
-        .filter_map(|var_id| program.state_address_for_source(root_instance_id, var_id))
-        .collect()
-}
-
-fn collect_statement_reads(
-    stmts: &[Statement],
-    funcs: &fxhash::FxHashMap<VarId, Function>,
-    active_functions: &mut crate::HashSet<VarId>,
-    reads: &mut crate::HashSet<VarId>,
-) {
-    for stmt in stmts {
-        match stmt {
-            Statement::Assign(assign) => {
-                collect_expression_reads(&assign.expr, funcs, active_functions, reads);
-                for dst in &assign.dst {
-                    for expr in &dst.index.0 {
-                        collect_expression_reads(expr, funcs, active_functions, reads);
-                    }
-                    collect_select_reads(&dst.select, funcs, active_functions, reads);
-                }
-            }
-            Statement::If(stmt) => {
-                collect_expression_reads(&stmt.cond, funcs, active_functions, reads);
-                collect_statement_reads(&stmt.true_side, funcs, active_functions, reads);
-                collect_statement_reads(&stmt.false_side, funcs, active_functions, reads);
-            }
-            Statement::IfReset(stmt) => {
-                collect_statement_reads(&stmt.true_side, funcs, active_functions, reads);
-                collect_statement_reads(&stmt.false_side, funcs, active_functions, reads);
-            }
-            Statement::Case(stmt) => {
-                collect_expression_reads(&stmt.case_target, funcs, active_functions, reads);
-                for arm in &stmt.arms {
-                    for pattern in &arm.patterns {
-                        match pattern {
-                            CasePattern::Eq(expr) => {
-                                collect_expression_reads(expr, funcs, active_functions, reads)
-                            }
-                            CasePattern::Range { lo, hi, .. } => {
-                                collect_expression_reads(lo, funcs, active_functions, reads);
-                                collect_expression_reads(hi, funcs, active_functions, reads);
-                            }
-                        }
-                    }
-                    collect_statement_reads(&arm.body, funcs, active_functions, reads);
-                }
-                collect_statement_reads(&stmt.default, funcs, active_functions, reads);
-            }
-            Statement::For(stmt) => {
-                collect_for_bound_reads(&stmt.range, funcs, active_functions, reads);
-                collect_statement_reads(&stmt.body, funcs, active_functions, reads);
-            }
-            Statement::SystemFunctionCall(call) => {
-                collect_system_function_reads(&call.kind, funcs, active_functions, reads);
-            }
-            Statement::FunctionCall(call) => {
-                collect_function_call_reads(call, funcs, active_functions, reads);
-            }
-            Statement::TbMethodCall(call) => match &call.method {
-                TbMethod::ClockNext { count, period } => {
-                    if let Some(expr) = count {
-                        collect_expression_reads(expr, funcs, active_functions, reads);
-                    }
-                    if let Some(expr) = period {
-                        collect_expression_reads(expr, funcs, active_functions, reads);
-                    }
-                }
-                TbMethod::ResetAssert { duration, .. } => {
-                    if let Some(expr) = duration {
-                        collect_expression_reads(expr, funcs, active_functions, reads);
-                    }
-                }
-                TbMethod::FileOpen { name, .. } => {
-                    collect_expression_reads(&name.0, funcs, active_functions, reads);
-                }
-                TbMethod::FileWrite { args } => {
-                    for arg in args {
-                        collect_expression_reads(&arg.0, funcs, active_functions, reads);
-                    }
-                }
-                TbMethod::FileClose | TbMethod::FileFlush => {}
-            },
-            Statement::Break | Statement::Unsupported(_) | Statement::Null => {}
-        }
-    }
-}
-
-fn collect_for_bound_reads(
-    range: &ForRange,
-    funcs: &fxhash::FxHashMap<VarId, Function>,
-    active_functions: &mut crate::HashSet<VarId>,
-    reads: &mut crate::HashSet<VarId>,
-) {
-    let (start, end) = match range {
-        ForRange::Forward { start, end, .. }
-        | ForRange::Reverse { start, end, .. }
-        | ForRange::Stepped { start, end, .. } => (start, end),
-    };
-    for bound in [start, end] {
-        if let ForBound::Expression(expr) = bound {
-            collect_expression_reads(expr, funcs, active_functions, reads);
-        }
-    }
-}
-
-fn collect_expression_reads(
-    expr: &Expression,
-    funcs: &fxhash::FxHashMap<VarId, Function>,
-    active_functions: &mut crate::HashSet<VarId>,
-    reads: &mut crate::HashSet<VarId>,
-) {
-    match expr {
-        Expression::Term(factor) => match factor.as_ref() {
-            Factor::Variable(var_id, index, select, _) => {
-                reads.insert(*var_id);
-                for expr in &index.0 {
-                    collect_expression_reads(expr, funcs, active_functions, reads);
-                }
-                collect_select_reads(select, funcs, active_functions, reads);
-            }
-            Factor::SystemFunctionCall(call) => {
-                collect_system_function_reads(&call.kind, funcs, active_functions, reads);
-            }
-            Factor::FunctionCall(call) => {
-                collect_function_call_reads(call, funcs, active_functions, reads);
-            }
-            Factor::Value(_) | Factor::Anonymous(_) | Factor::Unknown(_) => {}
-        },
-        Expression::Unary(_, inner, _) => {
-            collect_expression_reads(inner, funcs, active_functions, reads);
-        }
-        Expression::Binary(lhs, _, rhs, _) => {
-            collect_expression_reads(lhs, funcs, active_functions, reads);
-            collect_expression_reads(rhs, funcs, active_functions, reads);
-        }
-        Expression::Ternary(cond, then_expr, else_expr, _) => {
-            collect_expression_reads(cond, funcs, active_functions, reads);
-            collect_expression_reads(then_expr, funcs, active_functions, reads);
-            collect_expression_reads(else_expr, funcs, active_functions, reads);
-        }
-        Expression::Concatenation(parts, _) => {
-            for (value, repeat) in parts {
-                collect_expression_reads(value, funcs, active_functions, reads);
-                if let Some(repeat) = repeat {
-                    collect_expression_reads(repeat, funcs, active_functions, reads);
-                }
-            }
-        }
-        Expression::ArrayLiteral(items, _) => {
-            for item in items {
-                match item {
-                    ArrayLiteralItem::Value(value, repeat) => {
-                        collect_expression_reads(value, funcs, active_functions, reads);
-                        if let Some(repeat) = repeat {
-                            collect_expression_reads(repeat, funcs, active_functions, reads);
-                        }
-                    }
-                    ArrayLiteralItem::Defaul(value) => {
-                        collect_expression_reads(value, funcs, active_functions, reads);
-                    }
-                }
-            }
-        }
-        Expression::StructConstructor(_, fields, _) => {
-            for (_, value) in fields {
-                collect_expression_reads(value, funcs, active_functions, reads);
-            }
-        }
-    }
-}
-
-fn collect_select_reads(
-    select: &veryl_analyzer::ir::VarSelect,
-    funcs: &fxhash::FxHashMap<VarId, Function>,
-    active_functions: &mut crate::HashSet<VarId>,
-    reads: &mut crate::HashSet<VarId>,
-) {
-    for expr in &select.0 {
-        collect_expression_reads(expr, funcs, active_functions, reads);
-    }
-    if let Some((_, expr)) = &select.1 {
-        collect_expression_reads(expr, funcs, active_functions, reads);
-    }
-}
-
-fn collect_system_function_reads(
-    kind: &SystemFunctionKind,
-    funcs: &fxhash::FxHashMap<VarId, Function>,
-    active_functions: &mut crate::HashSet<VarId>,
-    reads: &mut crate::HashSet<VarId>,
-) {
-    let mut collect_input = |input: &SystemFunctionInput| {
-        collect_expression_reads(&input.0, funcs, active_functions, reads);
-    };
-    match kind {
-        SystemFunctionKind::Bits(input)
-        | SystemFunctionKind::Size(input)
-        | SystemFunctionKind::Clog2(input)
-        | SystemFunctionKind::Onehot(input)
-        | SystemFunctionKind::Signed(input)
-        | SystemFunctionKind::Unsigned(input) => collect_input(input),
-        SystemFunctionKind::Readmemh(input, _) => collect_input(input),
-        SystemFunctionKind::Display(inputs) | SystemFunctionKind::Write(inputs) => {
-            for input in inputs {
-                collect_input(input);
-            }
-        }
-        SystemFunctionKind::Assert { cond, args, .. } => {
-            collect_input(cond);
-            for input in args {
-                collect_input(input);
-            }
-        }
-        SystemFunctionKind::Finish => {}
-    }
-}
-
-fn collect_function_call_reads(
-    call: &FunctionCall,
-    funcs: &fxhash::FxHashMap<VarId, Function>,
-    active_functions: &mut crate::HashSet<VarId>,
-    reads: &mut crate::HashSet<VarId>,
-) {
-    for expr in call.inputs.values() {
-        collect_expression_reads(expr, funcs, active_functions, reads);
-    }
-    for destinations in call.outputs.values() {
-        for dst in destinations {
-            for expr in &dst.index.0 {
-                collect_expression_reads(expr, funcs, active_functions, reads);
-            }
-            collect_select_reads(&dst.select, funcs, active_functions, reads);
-        }
-    }
-    if active_functions.insert(call.id) {
-        if let Some(body) = function_body(funcs, call) {
-            collect_statement_reads(&body.statements, funcs, active_functions, reads);
-        }
-        active_functions.remove(&call.id);
-    }
-}
-
-fn compile_assert_message(
-    args: &[SystemFunctionInput],
-    ec: &ExprCompiler<'_>,
-) -> Option<GenericAssertMessage<SemanticArgument<AbsoluteAddr>>> {
-    if args.is_empty() {
-        return None;
-    }
-    if let Some(template) = static_string_expr(&args[0].0) {
-        let compiled_args = args[1..]
-            .iter()
-            .map(|arg| compile_assert_arg(arg, ec))
-            .collect::<Vec<_>>();
-        Some(GenericAssertMessage::Formatted {
-            template,
-            args: compiled_args,
+        .filter_map(|var_id| {
+            program
+                .frontend
+                .root_variable(var_id)
+                .map(|(address, _)| address)
         })
-    } else {
-        Some(GenericAssertMessage::DynamicArgs(
-            args.iter().map(|arg| compile_assert_arg(arg, ec)).collect(),
-        ))
-    }
+        .collect();
 }
 
 fn format_assert_arg(arg: &CompiledAssertArg, memory: *mut u8, spec: Option<char>) -> String {
@@ -910,32 +164,6 @@ fn render_assert_message(
     }
 }
 
-/// # Safety
-/// `ptr` must be valid for `byte_size` bytes of read access.
-#[inline(always)]
-unsafe fn read_le_u64(ptr: *const u8, byte_size: usize) -> u64 {
-    let mut buf = [0u8; 8];
-    unsafe {
-        std::ptr::copy_nonoverlapping(ptr, buf.as_mut_ptr(), byte_size.min(8));
-    }
-    u64::from_le_bytes(buf)
-}
-
-/// # Safety
-/// `ptr` must be valid for `byte_size` bytes of read access.
-unsafe fn read_le_wide(ptr: *const u8, byte_size: usize, width: usize) -> BigUint {
-    let mut buf = vec![0u8; byte_size];
-    unsafe {
-        std::ptr::copy_nonoverlapping(ptr, buf.as_mut_ptr(), byte_size);
-    }
-    let mut val = BigUint::from_bytes_le(&buf);
-    let extra_bits = byte_size * 8 - width;
-    if extra_bits > 0 {
-        val &= (BigUint::from(1u32) << width) - BigUint::from(1u32);
-    }
-    val
-}
-
 fn sim_set_u64<B: SimBackend>(sim: &mut crate::Simulator<B>, sig: SignalRef, value: u64) {
     match sig.width {
         0..=8 => sim.set(sig, value as u8),
@@ -946,1692 +174,7 @@ fn sim_set_u64<B: SimBackend>(sim: &mut crate::Simulator<B>, sig: SignalRef, val
     }
 }
 
-// ── Typed evaluation ───────────────────────────────────────────────────
-
-/// Binary operation on `TbValue`.  When both operands are `U64` the fast
-/// path runs entirely in registers; otherwise we promote to `BigUint`.
-#[inline]
-fn eval_binop(l: TbValue, op: Op, r: TbValue) -> TbValue {
-    match (&l, &r) {
-        (TbValue::U64(lv), TbValue::U64(rv)) => TbValue::U64(eval_binop_u64(*lv, op, *rv)),
-        _ => {
-            let lv = l.to_biguint();
-            let rv = r.to_biguint();
-            // Comparison / logic ops always return u64
-            match op {
-                Op::Eq
-                | Op::Ne
-                | Op::Less
-                | Op::LessEq
-                | Op::Greater
-                | Op::GreaterEq
-                | Op::LogicAnd
-                | Op::LogicOr => TbValue::U64(eval_binop_wide_cmp(&lv, op, &rv)),
-                _ => TbValue::Wide(eval_binop_wide(lv, op, rv)),
-            }
-        }
-    }
-}
-
-fn width_mask(width: usize) -> BigUint {
-    if width == 0 {
-        BigUint::ZERO
-    } else {
-        (BigUint::from(1u8) << width) - BigUint::from(1u8)
-    }
-}
-
-#[inline]
-fn width_mask_u64(width: usize) -> u64 {
-    match width {
-        0 => 0,
-        1..=63 => (1u64 << width) - 1,
-        _ => u64::MAX,
-    }
-}
-
-#[inline]
-fn signed_i128(value: u64, width: usize) -> i128 {
-    let value = value & width_mask_u64(width);
-    if width == 0 || width >= 64 {
-        (value as i64) as i128
-    } else if value & (1u64 << (width - 1)) == 0 {
-        value as i128
-    } else {
-        value as i128 - (1i128 << width)
-    }
-}
-
-fn normalized_bits(value: &TbValue, width: usize) -> BigUint {
-    value.to_biguint() & width_mask(width)
-}
-
-fn tb_value_from_bits(value: BigUint, width: usize) -> TbValue {
-    let value = value & width_mask(width);
-    if width <= 64 {
-        TbValue::U64(value.to_u64().unwrap_or(0))
-    } else {
-        TbValue::Wide(value)
-    }
-}
-
-fn signed_bigint(value: &TbValue, width: usize) -> BigInt {
-    let raw = normalized_bits(value, width);
-    if width == 0 || !raw.bit((width - 1) as u64) {
-        BigInt::from(raw)
-    } else {
-        BigInt::from(raw) - (BigInt::from(1u8) << width)
-    }
-}
-
-fn signed_bits(value: BigInt, width: usize) -> BigUint {
-    if width == 0 {
-        return BigUint::ZERO;
-    }
-    let modulus = BigUint::from(1u8) << width;
-    match value.sign() {
-        Sign::Minus => {
-            let magnitude = (-value).to_biguint().unwrap_or_default() % &modulus;
-            if magnitude == BigUint::ZERO {
-                BigUint::ZERO
-            } else {
-                modulus - magnitude
-            }
-        }
-        _ => value.to_biguint().unwrap_or_default() % modulus,
-    }
-}
-
-fn resize_tb_value(
-    value: &TbValue,
-    source_width: usize,
-    target_width: usize,
-    signed: bool,
-) -> TbValue {
-    if target_width == 0 {
-        return TbValue::U64(0);
-    }
-    if source_width == 0 {
-        let fill = if value.to_u64() & 1 == 0 {
-            BigUint::ZERO
-        } else {
-            width_mask(target_width)
-        };
-        return tb_value_from_bits(fill, target_width);
-    }
-
-    if let TbValue::U64(value) = value
-        && source_width <= 64
-        && target_width <= 64
-    {
-        let mut value = value & width_mask_u64(source_width);
-        if target_width > source_width && signed && value & (1u64 << (source_width - 1)) != 0 {
-            value |= width_mask_u64(target_width) ^ width_mask_u64(source_width);
-        }
-        return TbValue::U64(value & width_mask_u64(target_width));
-    }
-
-    let mut value = normalized_bits(value, source_width);
-    if target_width > source_width && signed && value.bit((source_width - 1) as u64) {
-        value |= width_mask(target_width) ^ width_mask(source_width);
-    }
-    tb_value_from_bits(value, target_width)
-}
-
-fn eval_typed_binop_u64(
-    l: u64,
-    op: Op,
-    r: u64,
-    lhs_width: usize,
-    rhs_width: usize,
-    result_width: usize,
-    lhs_signed: bool,
-    rhs_signed: bool,
-) -> u64 {
-    let l = l & width_mask_u64(lhs_width);
-    let r = r & width_mask_u64(rhs_width);
-    let result_mask = width_mask_u64(result_width);
-    let signed = lhs_signed && rhs_signed;
-    let bool_value = |value: bool| u64::from(value);
-
-    match op {
-        Op::Eq | Op::EqWildcard => bool_value(l == r),
-        Op::Ne | Op::NeWildcard => bool_value(l != r),
-        Op::Less if signed => bool_value(signed_i128(l, lhs_width) < signed_i128(r, rhs_width)),
-        Op::Less => bool_value(l < r),
-        Op::LessEq if signed => bool_value(signed_i128(l, lhs_width) <= signed_i128(r, rhs_width)),
-        Op::LessEq => bool_value(l <= r),
-        Op::Greater if signed => bool_value(signed_i128(l, lhs_width) > signed_i128(r, rhs_width)),
-        Op::Greater => bool_value(l > r),
-        Op::GreaterEq if signed => {
-            bool_value(signed_i128(l, lhs_width) >= signed_i128(r, rhs_width))
-        }
-        Op::GreaterEq => bool_value(l >= r),
-        Op::LogicAnd => bool_value(l != 0 && r != 0),
-        Op::LogicOr => bool_value(l != 0 || r != 0),
-        Op::Add => l.wrapping_add(r) & result_mask,
-        Op::Sub => l.wrapping_sub(r) & result_mask,
-        Op::Mul => l.wrapping_mul(r) & result_mask,
-        Op::Div if signed => {
-            let divisor = signed_i128(r, rhs_width);
-            if divisor == 0 {
-                0
-            } else {
-                (signed_i128(l, lhs_width) / divisor) as u64 & result_mask
-            }
-        }
-        Op::Div => l.checked_div(r).unwrap_or(0) & result_mask,
-        Op::Rem if signed => {
-            let divisor = signed_i128(r, rhs_width);
-            if divisor == 0 {
-                0
-            } else {
-                (signed_i128(l, lhs_width) % divisor) as u64 & result_mask
-            }
-        }
-        Op::Rem => l.checked_rem(r).unwrap_or(0) & result_mask,
-        Op::Pow => {
-            let mut exponent = r;
-            let mut base = l & result_mask;
-            let mut value = 1u64 & result_mask;
-            while exponent != 0 {
-                if exponent & 1 != 0 {
-                    value = ((value as u128 * base as u128) as u64) & result_mask;
-                }
-                exponent >>= 1;
-                if exponent != 0 {
-                    base = ((base as u128 * base as u128) as u64) & result_mask;
-                }
-            }
-            value
-        }
-        Op::BitAnd => (l & r) & result_mask,
-        Op::BitOr => (l | r) & result_mask,
-        Op::BitXor => (l ^ r) & result_mask,
-        Op::BitXnor => (!(l ^ r)) & result_mask,
-        Op::BitNand => (!(l & r)) & result_mask,
-        Op::BitNor => (!(l | r)) & result_mask,
-        Op::LogicShiftL | Op::ArithShiftL => {
-            if r >= result_width as u64 {
-                0
-            } else {
-                l.wrapping_shl(r as u32) & result_mask
-            }
-        }
-        Op::LogicShiftR => {
-            if r >= result_width as u64 {
-                0
-            } else {
-                (l >> r) & result_mask
-            }
-        }
-        Op::ArithShiftR if lhs_signed => {
-            let value = signed_i128(l, lhs_width);
-            if r >= result_width as u64 {
-                if value < 0 { result_mask } else { 0 }
-            } else {
-                ((value >> r) as u64) & result_mask
-            }
-        }
-        Op::ArithShiftR => {
-            if r >= result_width as u64 {
-                0
-            } else {
-                (l >> r) & result_mask
-            }
-        }
-        _ => unreachable!("operator is not a source-language binary op: {op:?}"),
-    }
-}
-
-fn eval_typed_binop(
-    l: TbValue,
-    op: Op,
-    r: TbValue,
-    lhs_width: usize,
-    rhs_width: usize,
-    result_width: usize,
-    lhs_signed: bool,
-    rhs_signed: bool,
-) -> TbValue {
-    if let (TbValue::U64(l), TbValue::U64(r)) = (&l, &r)
-        && lhs_width <= 64
-        && rhs_width <= 64
-        && result_width <= 64
-    {
-        return TbValue::U64(eval_typed_binop_u64(
-            *l,
-            op,
-            *r,
-            lhs_width,
-            rhs_width,
-            result_width,
-            lhs_signed,
-            rhs_signed,
-        ));
-    }
-    let lb = normalized_bits(&l, lhs_width);
-    let rb = normalized_bits(&r, rhs_width);
-    let signed = lhs_signed && rhs_signed;
-
-    let comparison = |value: bool| TbValue::U64(u64::from(value));
-    match op {
-        Op::Eq | Op::EqWildcard => comparison(lb == rb),
-        Op::Ne | Op::NeWildcard => comparison(lb != rb),
-        Op::Less if signed => {
-            comparison(signed_bigint(&l, lhs_width) < signed_bigint(&r, rhs_width))
-        }
-        Op::Less => comparison(lb < rb),
-        Op::LessEq if signed => {
-            comparison(signed_bigint(&l, lhs_width) <= signed_bigint(&r, rhs_width))
-        }
-        Op::LessEq => comparison(lb <= rb),
-        Op::Greater if signed => {
-            comparison(signed_bigint(&l, lhs_width) > signed_bigint(&r, rhs_width))
-        }
-        Op::Greater => comparison(lb > rb),
-        Op::GreaterEq if signed => {
-            comparison(signed_bigint(&l, lhs_width) >= signed_bigint(&r, rhs_width))
-        }
-        Op::GreaterEq => comparison(lb >= rb),
-        Op::LogicAnd => comparison(lb != BigUint::ZERO && rb != BigUint::ZERO),
-        Op::LogicOr => comparison(lb != BigUint::ZERO || rb != BigUint::ZERO),
-        Op::Add => tb_value_from_bits(lb + rb, result_width),
-        Op::Sub => tb_value_from_bits(
-            signed_bits(BigInt::from(lb) - BigInt::from(rb), result_width),
-            result_width,
-        ),
-        Op::Mul => tb_value_from_bits(lb * rb, result_width),
-        Op::Div if signed => {
-            let divisor = signed_bigint(&r, rhs_width);
-            if divisor == BigInt::from(0u8) {
-                TbValue::U64(0)
-            } else {
-                let quotient = signed_bigint(&l, lhs_width) / divisor;
-                tb_value_from_bits(signed_bits(quotient, result_width), result_width)
-            }
-        }
-        Op::Div => {
-            if rb == BigUint::ZERO {
-                TbValue::U64(0)
-            } else {
-                tb_value_from_bits(lb / rb, result_width)
-            }
-        }
-        Op::Rem if signed => {
-            let divisor = signed_bigint(&r, rhs_width);
-            if divisor == BigInt::from(0u8) {
-                TbValue::U64(0)
-            } else {
-                let remainder = signed_bigint(&l, lhs_width) % divisor;
-                tb_value_from_bits(signed_bits(remainder, result_width), result_width)
-            }
-        }
-        Op::Rem => {
-            if rb == BigUint::ZERO {
-                TbValue::U64(0)
-            } else {
-                tb_value_from_bits(lb % rb, result_width)
-            }
-        }
-        Op::Pow => {
-            if result_width == 0 {
-                TbValue::U64(0)
-            } else {
-                let modulus = BigUint::from(1u8) << result_width;
-                tb_value_from_bits(lb.modpow(&rb, &modulus), result_width)
-            }
-        }
-        Op::BitAnd => tb_value_from_bits(lb & rb, result_width),
-        Op::BitOr => tb_value_from_bits(lb | rb, result_width),
-        Op::BitXor => tb_value_from_bits(lb ^ rb, result_width),
-        Op::BitXnor => tb_value_from_bits((lb ^ rb) ^ width_mask(result_width), result_width),
-        Op::LogicShiftL | Op::ArithShiftL => {
-            let shift = rb.to_usize().unwrap_or(usize::MAX);
-            if shift >= result_width {
-                TbValue::U64(0)
-            } else {
-                tb_value_from_bits(lb << shift, result_width)
-            }
-        }
-        Op::LogicShiftR => {
-            let shift = rb.to_usize().unwrap_or(usize::MAX);
-            if shift >= result_width {
-                TbValue::U64(0)
-            } else {
-                tb_value_from_bits(lb >> shift, result_width)
-            }
-        }
-        Op::ArithShiftR if lhs_signed => {
-            let shift = rb.to_usize().unwrap_or(usize::MAX);
-            let value = signed_bigint(&l, lhs_width);
-            let shifted = if shift >= result_width {
-                if value.sign() == Sign::Minus {
-                    BigInt::from(-1)
-                } else {
-                    BigInt::from(0)
-                }
-            } else {
-                value >> shift
-            };
-            tb_value_from_bits(signed_bits(shifted, result_width), result_width)
-        }
-        Op::ArithShiftR => {
-            let shift = rb.to_usize().unwrap_or(usize::MAX);
-            if shift >= result_width {
-                TbValue::U64(0)
-            } else {
-                tb_value_from_bits(lb >> shift, result_width)
-            }
-        }
-        Op::BitNand => tb_value_from_bits((lb & rb) ^ width_mask(result_width), result_width),
-        Op::BitNor => tb_value_from_bits((lb | rb) ^ width_mask(result_width), result_width),
-        _ => unreachable!("operator is not a source-language binary op: {op:?}"),
-    }
-}
-
-fn eval_typed_unop(op: Op, value: &TbValue, operand_width: usize, result_width: usize) -> TbValue {
-    if let TbValue::U64(value) = value
-        && operand_width <= 64
-        && result_width <= 64
-    {
-        let bits = value & width_mask_u64(operand_width);
-        let value = match op {
-            Op::LogicNot => u64::from(bits == 0),
-            Op::BitAnd => u64::from(bits == width_mask_u64(operand_width)),
-            Op::BitNand => u64::from(bits != width_mask_u64(operand_width)),
-            Op::BitOr => u64::from(bits != 0),
-            Op::BitNor => u64::from(bits == 0),
-            Op::BitXor => u64::from(!bits.count_ones().is_multiple_of(2)),
-            Op::BitXnor => u64::from(bits.count_ones().is_multiple_of(2)),
-            Op::Add => bits & width_mask_u64(result_width),
-            Op::Sub => bits.wrapping_neg() & width_mask_u64(result_width),
-            Op::BitNot => !bits & width_mask_u64(result_width),
-            _ => unreachable!("operator is not a source-language unary op: {op:?}"),
-        };
-        return TbValue::U64(value);
-    }
-
-    let bits = normalized_bits(value, operand_width);
-    let reduced = match op {
-        Op::LogicNot => Some(bits == BigUint::ZERO),
-        Op::BitAnd => Some(bits == width_mask(operand_width)),
-        Op::BitNand => Some(bits != width_mask(operand_width)),
-        Op::BitOr => Some(bits != BigUint::ZERO),
-        Op::BitNor => Some(bits == BigUint::ZERO),
-        Op::BitXor | Op::BitXnor => {
-            let odd = bits.iter_u64_digits().map(u64::count_ones).sum::<u32>() % 2 != 0;
-            Some(if matches!(op, Op::BitXor) { odd } else { !odd })
-        }
-        _ => None,
-    };
-    if let Some(value) = reduced {
-        return TbValue::U64(u64::from(value));
-    }
-
-    match op {
-        Op::Add => tb_value_from_bits(bits, result_width),
-        Op::Sub => tb_value_from_bits(signed_bits(-BigInt::from(bits), result_width), result_width),
-        Op::BitNot => tb_value_from_bits(bits ^ width_mask(operand_width), result_width),
-        _ => unreachable!("operator is not a source-language unary op: {op:?}"),
-    }
-}
-
-#[inline]
-fn eval_binop_u64(l: u64, op: Op, r: u64) -> u64 {
-    match op {
-        Op::Add => l.wrapping_add(r),
-        Op::Sub => l.wrapping_sub(r),
-        Op::Mul => l.wrapping_mul(r),
-        Op::Div => l.checked_div(r).unwrap_or(0),
-        Op::Rem => l.checked_rem(r).unwrap_or(0),
-        Op::BitAnd => l & r,
-        Op::BitOr => l | r,
-        Op::BitXor => l ^ r,
-        Op::LogicShiftL => {
-            if r >= 64 {
-                0
-            } else {
-                l << r
-            }
-        }
-        Op::LogicShiftR => {
-            if r >= 64 {
-                0
-            } else {
-                l >> r
-            }
-        }
-        Op::ArithShiftL => {
-            if r >= 64 {
-                0
-            } else {
-                l << r
-            }
-        }
-        Op::ArithShiftR => {
-            if r >= 64 {
-                ((l as i64) >> 63) as u64
-            } else {
-                ((l as i64) >> r) as u64
-            }
-        }
-        Op::Eq => (l == r) as u64,
-        Op::Ne => (l != r) as u64,
-        Op::Less => (l < r) as u64,
-        Op::LessEq => (l <= r) as u64,
-        Op::Greater => (l > r) as u64,
-        Op::GreaterEq => (l >= r) as u64,
-        Op::LogicAnd => ((l != 0) && (r != 0)) as u64,
-        Op::LogicOr => ((l != 0) || (r != 0)) as u64,
-        _ => unreachable!("operator is not testbench bytecode plumbing: {op:?}"),
-    }
-}
-
-fn eval_binop_wide(l: BigUint, op: Op, r: BigUint) -> BigUint {
-    match op {
-        Op::Add => l + r,
-        Op::Sub => {
-            if l >= r {
-                l - r
-            } else {
-                BigUint::ZERO
-            }
-        }
-        Op::Mul => l * r,
-        Op::Div => {
-            if r == BigUint::ZERO {
-                BigUint::ZERO
-            } else {
-                l / r
-            }
-        }
-        Op::Rem => {
-            if r == BigUint::ZERO {
-                BigUint::ZERO
-            } else {
-                l % r
-            }
-        }
-        Op::BitAnd => l & r,
-        Op::BitOr => l | r,
-        Op::BitXor => l ^ r,
-        Op::LogicShiftL => {
-            let s: u64 = (&r).try_into().unwrap_or(256);
-            l << s
-        }
-        Op::LogicShiftR => {
-            let s: u64 = (&r).try_into().unwrap_or(256);
-            l >> s
-        }
-        _ => unreachable!("operator is not wide testbench bytecode plumbing: {op:?}"),
-    }
-}
-
-fn eval_binop_wide_cmp(l: &BigUint, op: Op, r: &BigUint) -> u64 {
-    match op {
-        Op::Eq => (l == r) as u64,
-        Op::Ne => (l != r) as u64,
-        Op::Less => (l < r) as u64,
-        Op::LessEq => (l <= r) as u64,
-        Op::Greater => (l > r) as u64,
-        Op::GreaterEq => (l >= r) as u64,
-        Op::LogicAnd => ((*l != BigUint::ZERO) && (*r != BigUint::ZERO)) as u64,
-        Op::LogicOr => ((*l != BigUint::ZERO) || (*r != BigUint::ZERO)) as u64,
-        _ => unreachable!("operator is not testbench comparison plumbing: {op:?}"),
-    }
-}
-
 // ── Expression compiler ────────────────────────────────────────────────
-
-fn lower_testbench_operator(op: VerylOp) -> Op {
-    match op {
-        VerylOp::Add => Op::Add,
-        VerylOp::Sub => Op::Sub,
-        VerylOp::Mul => Op::Mul,
-        VerylOp::Div => Op::Div,
-        VerylOp::Rem => Op::Rem,
-        VerylOp::Pow => Op::Pow,
-        VerylOp::BitAnd => Op::BitAnd,
-        VerylOp::BitOr => Op::BitOr,
-        VerylOp::BitXor => Op::BitXor,
-        VerylOp::BitXnor => Op::BitXnor,
-        VerylOp::BitNand => Op::BitNand,
-        VerylOp::BitNor => Op::BitNor,
-        VerylOp::LogicShiftL => Op::LogicShiftL,
-        VerylOp::LogicShiftR => Op::LogicShiftR,
-        VerylOp::ArithShiftL => Op::ArithShiftL,
-        VerylOp::ArithShiftR => Op::ArithShiftR,
-        VerylOp::Eq => Op::Eq,
-        VerylOp::EqWildcard => Op::EqWildcard,
-        VerylOp::Ne => Op::Ne,
-        VerylOp::NeWildcard => Op::NeWildcard,
-        VerylOp::Less => Op::Less,
-        VerylOp::LessEq => Op::LessEq,
-        VerylOp::Greater => Op::Greater,
-        VerylOp::GreaterEq => Op::GreaterEq,
-        VerylOp::LogicAnd => Op::LogicAnd,
-        VerylOp::LogicOr => Op::LogicOr,
-        VerylOp::LogicNot => Op::LogicNot,
-        VerylOp::BitNot => Op::BitNot,
-        _ => unreachable!("operator cannot be represented in testbench bytecode: {op:?}"),
-    }
-}
-
-struct ExprCompiler<'a> {
-    program: &'a Program,
-    testbench_source: &'a VerylTestbenchSource,
-    /// Root module instance ID, cached for repeated lookups.
-    root_instance_id: crate::ir::InstanceId,
-    root_module_id: crate::ir::ModuleId,
-}
-
-impl ExprCompiler<'_> {
-    fn compile(&self, expr: &Expression) -> ExprBytecode<StateLocation<AbsoluteAddr>> {
-        let mut ops = Vec::new();
-        self.emit(expr, &mut ops);
-        ExprBytecode::new(ops)
-    }
-
-    fn compile_with_width(
-        &self,
-        expr: &Expression,
-        width: usize,
-    ) -> ExprBytecode<StateLocation<AbsoluteAddr>> {
-        let mut ops = Vec::new();
-        self.emit_in_context(
-            expr,
-            &mut ops,
-            Some(ValueContext {
-                width,
-                signed: expression_signed(expr),
-            }),
-        );
-        ExprBytecode::new(ops)
-    }
-
-    fn natural_width(&self, expr: &Expression) -> usize {
-        get_expr_width(expr)
-            .filter(|width| *width != 0)
-            .unwrap_or_else(|| self.infer_expr_width(expr).max(1))
-    }
-
-    fn root_context(&self, expr: &Expression) -> ValueContext {
-        ValueContext {
-            width: self
-                .natural_width(expr)
-                .max(expr.comptime().expr_context.width),
-            signed: expression_signed(expr),
-        }
-    }
-
-    fn emit(&self, expr: &Expression, ops: &mut Vec<UnboundTbOpcode>) {
-        self.emit_in_context(expr, ops, Some(self.root_context(expr)));
-    }
-
-    fn resize_result(
-        &self,
-        ops: &mut Vec<UnboundTbOpcode>,
-        source: ValueContext,
-        target: Option<ValueContext>,
-    ) -> ValueContext {
-        let Some(target) = target else {
-            return source;
-        };
-        if source.width != target.width {
-            ops.push(TbOpcode::Resize {
-                source_width: source.width,
-                target_width: target.width,
-                signed: target.signed,
-            });
-        }
-        target
-    }
-
-    fn emit_in_context(
-        &self,
-        expr: &Expression,
-        ops: &mut Vec<UnboundTbOpcode>,
-        context: Option<ValueContext>,
-    ) -> ValueContext {
-        match expr {
-            Expression::Term(f) => {
-                self.emit_factor(f, ops);
-                let source = ValueContext {
-                    width: self.natural_width(expr),
-                    signed: expression_signed(expr),
-                };
-                self.resize_result(ops, source, context)
-            }
-            Expression::Unary(op, inner, _) => {
-                let reduction = matches!(
-                    op,
-                    VerylOp::BitAnd
-                        | VerylOp::BitNand
-                        | VerylOp::BitOr
-                        | VerylOp::BitNor
-                        | VerylOp::BitXor
-                        | VerylOp::BitXnor
-                        | VerylOp::LogicNot
-                );
-                let natural_width = self.natural_width(inner);
-                let operand_context = if reduction {
-                    None
-                } else {
-                    Some(ValueContext {
-                        width: natural_width.max(context.map(|x| x.width).unwrap_or(0)),
-                        signed: context
-                            .map(|x| x.signed)
-                            .unwrap_or_else(|| expression_signed(inner)),
-                    })
-                };
-                let operand = self.emit_in_context(inner, ops, operand_context);
-                let result_width = if reduction { 1 } else { operand.width };
-                match op {
-                    VerylOp::Add
-                    | VerylOp::Sub
-                    | VerylOp::BitNot
-                    | VerylOp::BitAnd
-                    | VerylOp::BitNand
-                    | VerylOp::BitOr
-                    | VerylOp::BitNor
-                    | VerylOp::BitXor
-                    | VerylOp::BitXnor
-                    | VerylOp::LogicNot => ops.push(TbOpcode::TypedUnary {
-                        op: lower_testbench_operator(*op),
-                        operand_width: operand.width,
-                        result_width,
-                    }),
-                    _ => unreachable!("operator is not unary in a testbench expression: {op:?}"),
-                }
-                let result = ValueContext {
-                    width: result_width,
-                    signed: if reduction { false } else { operand.signed },
-                };
-                self.resize_result(ops, result, context)
-            }
-            Expression::Binary(lhs, op, rhs, _) => {
-                if matches!(op, VerylOp::As) {
-                    let cast = cast_semantics(lhs, rhs)
-                        .expect("analyzed testbench cast must have a concrete target");
-                    let source = self.emit_in_context(
-                        lhs,
-                        ops,
-                        Some(ValueContext {
-                            width: cast.width,
-                            signed: cast.source_signed,
-                        }),
-                    );
-                    let casted = ValueContext {
-                        width: source.width,
-                        signed: cast.result_signed,
-                    };
-                    return self.resize_result(ops, casted, context);
-                }
-
-                let lhs_width = self.natural_width(lhs);
-                let rhs_width = self.natural_width(rhs);
-                let semantics = binary_semantics(
-                    *op,
-                    lhs_width,
-                    rhs_width,
-                    expression_signed(lhs),
-                    expression_signed(rhs),
-                    context,
-                );
-                let lhs = self.emit_in_context(lhs, ops, semantics.lhs_context);
-                let rhs = self.emit_in_context(rhs, ops, semantics.rhs_context);
-                match op {
-                    VerylOp::Pow
-                    | VerylOp::Div
-                    | VerylOp::Rem
-                    | VerylOp::Mul
-                    | VerylOp::Add
-                    | VerylOp::Sub
-                    | VerylOp::ArithShiftL
-                    | VerylOp::ArithShiftR
-                    | VerylOp::LogicShiftL
-                    | VerylOp::LogicShiftR
-                    | VerylOp::LessEq
-                    | VerylOp::GreaterEq
-                    | VerylOp::Less
-                    | VerylOp::Greater
-                    | VerylOp::Eq
-                    | VerylOp::EqWildcard
-                    | VerylOp::Ne
-                    | VerylOp::NeWildcard
-                    | VerylOp::BitAnd
-                    | VerylOp::BitOr
-                    | VerylOp::BitXor
-                    | VerylOp::BitXnor
-                    | VerylOp::LogicAnd
-                    | VerylOp::LogicOr => ops.push(TbOpcode::TypedBinOp {
-                        op: lower_testbench_operator(*op),
-                        lhs_width: lhs.width,
-                        rhs_width: rhs.width,
-                        result_width: semantics.result_width,
-                        lhs_signed: semantics.lhs_signed,
-                        rhs_signed: semantics.rhs_signed,
-                    }),
-                    _ => unreachable!("operator is not binary in a testbench expression: {op:?}"),
-                }
-                let result = ValueContext {
-                    width: semantics.result_width,
-                    signed: semantics.result_signed,
-                };
-                self.resize_result(ops, result, context)
-            }
-            Expression::Ternary(cond, then_expr, else_expr, _) => {
-                self.emit_in_context(cond, ops, None);
-                let branch_context = ValueContext {
-                    width: self
-                        .natural_width(then_expr)
-                        .max(self.natural_width(else_expr))
-                        .max(context.map(|x| x.width).unwrap_or(0)),
-                    signed: context.map(|x| x.signed).unwrap_or_else(|| {
-                        expression_signed(then_expr) && expression_signed(else_expr)
-                    }),
-                };
-                let mut then_ops = Vec::new();
-                self.emit_in_context(then_expr, &mut then_ops, Some(branch_context));
-                let mut else_ops = Vec::new();
-                self.emit_in_context(else_expr, &mut else_ops, Some(branch_context));
-                ops.push(TbOpcode::Ternary {
-                    then_len: then_ops.len(),
-                    else_len: else_ops.len(),
-                });
-                ops.extend(then_ops);
-                ops.extend(else_ops);
-                branch_context
-            }
-            Expression::Concatenation(parts, _) => {
-                ops.push(TbOpcode::ConstU64(0));
-                let mut result_width = 0usize;
-                for (val_expr, repeat_expr) in parts {
-                    let part_width = self.natural_width(val_expr);
-                    let repeat = repeat_expr
-                        .as_ref()
-                        .and_then(Self::try_const_usize)
-                        .unwrap_or(1);
-                    for _ in 0..repeat {
-                        self.emit_in_context(val_expr, ops, None);
-                        result_width = result_width.saturating_add(part_width);
-                        ops.push(TbOpcode::ConcatPart {
-                            part_width,
-                            result_width,
-                        });
-                    }
-                }
-                self.resize_result(
-                    ops,
-                    ValueContext {
-                        width: result_width.max(1),
-                        signed: false,
-                    },
-                    context,
-                )
-            }
-            Expression::ArrayLiteral(..) | Expression::StructConstructor(..) => {
-                unreachable!("aggregate testbench expressions must be lowered before bytecode")
-            }
-        }
-    }
-
-    fn emit_factor(&self, factor: &Factor, ops: &mut Vec<UnboundTbOpcode>) {
-        match factor {
-            Factor::Variable(var_id, index, select, comptime) => {
-                if comptime.is_const
-                    && let Ok(value) = comptime.get_value()
-                {
-                    self.emit_constant_value(value, ops);
-                } else if let Some(sig) = self.resolve_var(var_id) {
-                    self.emit_var_access(var_id, sig, index, select, ops);
-                } else if let Ok(value) = comptime.get_value() {
-                    self.emit_constant_value(value, ops);
-                } else {
-                    unreachable!("unresolved non-constant testbench variable {var_id:?}");
-                }
-            }
-            Factor::Value(comptime) => {
-                if let Ok(val) = comptime.get_value() {
-                    self.emit_constant_value(val, ops);
-                } else {
-                    unreachable!("analyzed testbench literal has no value");
-                }
-            }
-            Factor::FunctionCall(fc) => {
-                self.emit_function_call(fc, ops);
-            }
-            Factor::SystemFunctionCall(call) => match &call.kind {
-                SystemFunctionKind::Signed(input) | SystemFunctionKind::Unsigned(input) => {
-                    self.emit(&input.0, ops);
-                }
-                _ => {
-                    let value = call
-                        .comptime
-                        .get_value()
-                        .expect("testbench system function must be compile-time evaluable");
-                    self.emit_constant_value(value, ops);
-                }
-            },
-            Factor::Anonymous(comptime) | Factor::Unknown(comptime) => {
-                let value = comptime
-                    .get_value()
-                    .expect("resolved testbench factor must have a value");
-                self.emit_constant_value(value, ops);
-            }
-        }
-    }
-
-    fn emit_constant_value(
-        &self,
-        value: &veryl_analyzer::value::Value,
-        ops: &mut Vec<UnboundTbOpcode>,
-    ) {
-        if value.width() <= 64 {
-            ops.push(TbOpcode::ConstU64(value.payload_u64()));
-        } else {
-            ops.push(TbOpcode::ConstWide(value.payload().into_owned()));
-        }
-    }
-
-    /// Emit bytecode for a function call used as an expression value.
-    /// Inline-expands: store args → emit body assigns → load return value.
-    fn emit_function_call(
-        &self,
-        fc: &veryl_analyzer::ir::FunctionCall,
-        ops: &mut Vec<UnboundTbOpcode>,
-    ) {
-        let func = match self.testbench_source.functions.get(&fc.id) {
-            Some(f) => f,
-            None => {
-                ops.push(TbOpcode::ConstU64(0));
-                return;
-            }
-        };
-        let func_body = match if let Some(idx) = &fc.index {
-            func.get_function(idx)
-        } else {
-            func.get_function(&[])
-        } {
-            Some(fb) => fb,
-            None => {
-                ops.push(TbOpcode::ConstU64(0));
-                return;
-            }
-        };
-
-        // 1. Store input arguments into memory
-        for (arg_path, arg_expr) in &fc.inputs {
-            if let Some(&arg_var_id) = func_body.arg_map.get(arg_path) {
-                if let Some(sig) = self.resolve_var(&arg_var_id) {
-                    self.emit_in_context(
-                        arg_expr,
-                        ops,
-                        Some(ValueContext {
-                            width: sig.width,
-                            signed: expression_signed(arg_expr),
-                        }),
-                    );
-                    ops.push(TbOpcode::StoreU64 {
-                        location: self.state_location(arg_var_id, 0),
-                        byte_size: get_byte_size(sig.width),
-                    });
-                }
-            }
-        }
-
-        // 2. Emit body statements as bytecode (only Assign is supported)
-        for stmt in &func_body.statements {
-            if let veryl_analyzer::ir::Statement::Assign(a) = stmt {
-                if let Some(first_dst) = a.dst.first() {
-                    if let Some(dst_sig) = self.resolve_var(&first_dst.id) {
-                        self.emit_in_context(
-                            &a.expr,
-                            ops,
-                            Some(ValueContext {
-                                width: dst_sig.width,
-                                signed: expression_signed(&a.expr),
-                            }),
-                        );
-                        ops.push(TbOpcode::StoreU64 {
-                            location: self.state_location(first_dst.id, 0),
-                            byte_size: get_byte_size(dst_sig.width),
-                        });
-                    }
-                }
-            }
-            // Non-assign statements (if/for in function body) are skipped.
-            // This covers the common case of pure computation functions.
-        }
-
-        // 3. Load return value
-        if let Some(ret_var_id) = &func_body.ret {
-            if let Some(sig) = self.resolve_var(ret_var_id) {
-                self.emit_load(*ret_var_id, 0, sig.width, ops);
-            } else {
-                ops.push(TbOpcode::ConstU64(0));
-            }
-        } else {
-            ops.push(TbOpcode::ConstU64(0));
-        }
-    }
-
-    /// Emit bytecode for a variable access, handling static and dynamic
-    /// array indices and bit selects.
-    fn emit_var_access(
-        &self,
-        var_id: &VarId,
-        sig: SemanticSignal<AbsoluteAddr>,
-        index: &veryl_analyzer::ir::VarIndex,
-        select: &veryl_analyzer::ir::VarSelect,
-        ops: &mut Vec<UnboundTbOpcode>,
-    ) {
-        let p = self.program;
-        let info = match p
-            .frontend
-            .module_variables
-            .get(&self.root_module_id)
-            .and_then(|v| v.get(var_id))
-        {
-            Some(i) => i,
-            None => {
-                self.emit_load(*var_id, 0, sig.width, ops);
-                return;
-            }
-        };
-
-        // No index or select → whole variable
-        if index.0.is_empty() && select.0.is_empty() && select.1.is_none() {
-            self.emit_load(*var_id, 0, sig.width, ops);
-            return;
-        }
-
-        let array_total: usize = info.array_dims.iter().product::<usize>().max(1);
-        let element_width = info.width / array_total;
-
-        // Compute array strides
-        let mut strides_bits = vec![element_width; info.array_dims.len()];
-        if !info.array_dims.is_empty() {
-            let mut stride = element_width;
-            for i in (0..info.array_dims.len()).rev() {
-                strides_bits[i] = stride;
-                stride *= info.array_dims[i];
-            }
-        }
-
-        // Process unpacked array indices
-        let mut static_bit_offset: usize = 0;
-        let mut dynamic_emitted = false;
-
-        for (i, idx_expr) in index.0.iter().enumerate() {
-            if i >= info.array_dims.len() {
-                break;
-            }
-            let stride = strides_bits[i];
-
-            if let Some(idx_val) = Self::try_const_usize(idx_expr) {
-                // Static index: accumulate into offset
-                static_bit_offset += idx_val * stride;
-            } else {
-                // Dynamic index: emit the index expression, then LoadIndexed
-                let base_byte_offset = static_bit_offset / 8;
-                let stride_bytes = get_byte_size(stride);
-                let elem_byte_size = get_byte_size(element_width);
-                self.emit(idx_expr, ops);
-                ops.push(TbOpcode::LoadIndexed {
-                    location: self.state_location(*var_id, base_byte_offset),
-                    stride_bytes,
-                    element_byte_size: elem_byte_size,
-                    element_width,
-                });
-                dynamic_emitted = true;
-                // After a dynamic index, remaining indices would need chaining.
-                // For now, only single dynamic index is supported.
-                break;
-            }
-        }
-
-        if dynamic_emitted {
-            // Apply bit select on top of dynamic result if present
-            if select.1.is_some() || !select.0.is_empty() {
-                self.emit_post_select(select, element_width, ops);
-            }
-            return;
-        }
-
-        // All indices were static — apply bit select
-        let accessed_width = if index.0.len() >= info.array_dims.len() {
-            element_width
-        } else if index.0.is_empty() {
-            info.width
-        } else {
-            strides_bits[index.0.len() - 1]
-        };
-
-        if select.0.is_empty() && select.1.is_none() {
-            // No bit select, just load the element
-            let byte_offset = static_bit_offset / 8;
-            let sub = static_bit_offset % 8;
-            if sub == 0 {
-                self.emit_load(*var_id, byte_offset, accessed_width, ops);
-            } else {
-                let load_width = accessed_width + sub;
-                self.emit_load(*var_id, byte_offset, load_width, ops);
-                ops.push(TbOpcode::ConstU64(sub as u64));
-                ops.push(TbOpcode::BinOp(Op::LogicShiftR));
-                if accessed_width < 64 {
-                    ops.push(TbOpcode::ConstU64((1u64 << accessed_width) - 1));
-                    ops.push(TbOpcode::BinOp(Op::BitAnd));
-                }
-            }
-            return;
-        }
-
-        // Static bit select
-        let (sel_lsb, sel_width, is_dynamic_select) = self.resolve_select(select, ops);
-
-        if is_dynamic_select {
-            // Dynamic bit select: load full value, shift by dynamic amount, mask
-            let byte_offset = static_bit_offset / 8;
-            let total_byte_size = get_byte_size(accessed_width);
-            ops.push(TbOpcode::LoadBitSelect {
-                location: self.state_location(*var_id, byte_offset),
-                base_byte_size: total_byte_size,
-                select_width: sel_width,
-            });
-            return;
-        }
-
-        let bit_offset = static_bit_offset + sel_lsb;
-        let byte_offset = bit_offset / 8;
-        let sub = bit_offset % 8;
-        if sub == 0 {
-            self.emit_load(*var_id, byte_offset, sel_width, ops);
-        } else {
-            let load_width = sel_width + sub;
-            self.emit_load(*var_id, byte_offset, load_width, ops);
-            ops.push(TbOpcode::ConstU64(sub as u64));
-            ops.push(TbOpcode::BinOp(Op::LogicShiftR));
-            if sel_width < 64 {
-                ops.push(TbOpcode::ConstU64((1u64 << sel_width) - 1));
-                ops.push(TbOpcode::BinOp(Op::BitAnd));
-            }
-        }
-    }
-
-    /// Resolve a VarSelect to `(lsb, width, is_dynamic)`.
-    /// If any index is dynamic, emits the dynamic index expression to `ops`
-    /// and returns `is_dynamic = true`.
-    fn resolve_select(
-        &self,
-        select: &veryl_analyzer::ir::VarSelect,
-        ops: &mut Vec<UnboundTbOpcode>,
-    ) -> (usize, usize, bool) {
-        if let Some((op, range_expr)) = &select.1 {
-            let anchor_expr = select.0.last();
-            let anchor = anchor_expr.and_then(Self::try_const_usize);
-            let range_val = Self::try_const_usize(range_expr);
-
-            if let (Some(a), Some(v)) = (anchor, range_val) {
-                let (lsb, msb) = match op {
-                    veryl_analyzer::ir::VarSelectOp::Colon => (v, a),
-                    veryl_analyzer::ir::VarSelectOp::PlusColon => (a, a + v - 1),
-                    veryl_analyzer::ir::VarSelectOp::MinusColon => (a.saturating_sub(v) + 1, a),
-                    veryl_analyzer::ir::VarSelectOp::Step => (a * v, (a + 1) * v - 1),
-                };
-                return (lsb, msb - lsb + 1, false);
-            }
-
-            // Dynamic select: emit the anchor expression
-            if let Some(anchor_expr) = anchor_expr {
-                self.emit(anchor_expr, ops);
-            } else {
-                ops.push(TbOpcode::ConstU64(0));
-            }
-            let width = range_val.unwrap_or(1);
-            return (0, width, true);
-        }
-
-        // Simple bit index (no range)
-        if let Some(first) = select.0.first() {
-            if let Some(idx) = Self::try_const_usize(first) {
-                return (idx, 1, false);
-            }
-            // Dynamic single bit select
-            self.emit(first, ops);
-            return (0, 1, true);
-        }
-
-        (0, 0, false)
-    }
-
-    /// Emit post-load bit select operations on a value already on the stack
-    /// (for dynamic array element access followed by bit select).
-    fn emit_post_select(
-        &self,
-        select: &veryl_analyzer::ir::VarSelect,
-        _base_width: usize,
-        ops: &mut Vec<UnboundTbOpcode>,
-    ) {
-        let (lsb, width, is_dynamic) = self.resolve_select(select, ops);
-        if is_dynamic {
-            // Stack: [value, bit_index]
-            ops.push(TbOpcode::BinOp(Op::LogicShiftR));
-            if width < 64 {
-                ops.push(TbOpcode::ConstU64((1u64 << width) - 1));
-                ops.push(TbOpcode::BinOp(Op::BitAnd));
-            }
-        } else if lsb > 0 || width > 0 {
-            if lsb > 0 {
-                ops.push(TbOpcode::ConstU64(lsb as u64));
-                ops.push(TbOpcode::BinOp(Op::LogicShiftR));
-            }
-            if width > 0 && width < 64 {
-                ops.push(TbOpcode::ConstU64((1u64 << width) - 1));
-                ops.push(TbOpcode::BinOp(Op::BitAnd));
-            }
-        }
-    }
-
-    /// Emit a LoadU64 or LoadWide opcode for the given byte offset and bit width.
-    fn emit_load(
-        &self,
-        var_id: VarId,
-        byte_offset: usize,
-        width: usize,
-        ops: &mut Vec<UnboundTbOpcode>,
-    ) {
-        let byte_size = get_byte_size(width);
-        if byte_size <= 8 {
-            let mask = if width >= 64 {
-                u64::MAX
-            } else {
-                (1u64 << width) - 1
-            };
-            ops.push(TbOpcode::LoadU64 {
-                location: self.state_location(var_id, byte_offset),
-                byte_size,
-                mask,
-            });
-        } else {
-            ops.push(TbOpcode::LoadWide {
-                location: self.state_location(var_id, byte_offset),
-                byte_size,
-                width,
-            });
-        }
-    }
-
-    fn state_location(&self, var_id: VarId, byte_offset: usize) -> StateLocation<AbsoluteAddr> {
-        StateLocation {
-            address: self
-                .program
-                .state_address_for_source(self.root_instance_id, var_id)
-                .expect("frontend state projection is complete"),
-            byte_offset,
-        }
-    }
-
-    /// Resolve VarIndex (unpacked array) and VarSelect (bit select) to
-    /// a concrete (byte_offset, bit_width) pair.
-    ///
-    /// For static indices, adjusts the offset and narrows the width.
-    /// Dynamic indices are not supported and fall back to the full variable.
-    /// Infer the bit width of an expression. Falls back to comptime if available,
-    /// otherwise resolves from VariableInfo for variables.
-    fn infer_expr_width(&self, expr: &Expression) -> usize {
-        let ctx_width = expr.comptime().expr_context.width;
-        if ctx_width > 0 {
-            return ctx_width;
-        }
-        // Try type-level width
-        if let Some(w) = expr.comptime().r#type.total_width() {
-            if w > 0 {
-                return w;
-            }
-        }
-        // For terms, look up variable info
-        if let Expression::Term(f) = expr {
-            match f.as_ref() {
-                Factor::Variable(var_id, _, _, _) => {
-                    let p = self.program;
-                    if let Some(vars) = p.frontend.module_variables.get(&self.root_module_id) {
-                        if let Some(info) = vars.get(var_id) {
-                            return info.width;
-                        }
-                    }
-                }
-                Factor::Value(c) => {
-                    if let Ok(v) = c.get_value() {
-                        return v.width();
-                    }
-                }
-                _ => {}
-            }
-        }
-        0
-    }
-
-    fn try_const_usize(expr: &Expression) -> Option<usize> {
-        match expr {
-            Expression::Term(f) => match f.as_ref() {
-                Factor::Value(c) => c.get_value().ok().map(|v| v.payload_u64() as usize),
-                _ => None,
-            },
-            _ => None,
-        }
-    }
-
-    fn resolve_var(&self, var_id: &VarId) -> Option<SemanticSignal<AbsoluteAddr>> {
-        let p = self.program;
-        let vars = p.frontend.module_variables.get(&self.root_module_id)?;
-        let info = vars.get(var_id)?;
-        Some(SemanticSignal {
-            address: p.state_address_for_source(self.root_instance_id, *var_id)?,
-            width: info.width,
-        })
-    }
-}
-
-// ── Builder ────────────────────────────────────────────────────────────
-
-struct SemanticTestbenchBuilder<'a> {
-    program: &'a Program,
-    testbench_source: &'a VerylTestbenchSource,
-    event_map: std::collections::HashMap<StrId, AbsoluteAddr>,
-    signal_map: std::collections::HashMap<StrId, SemanticSignal<AbsoluteAddr>>,
-    default_reset_duration: u64,
-}
-
-impl<'a> SemanticTestbenchBuilder<'a> {
-    fn new(program: &'a Program, testbench_source: &'a VerylTestbenchSource) -> Self {
-        Self {
-            program,
-            testbench_source,
-            event_map: Default::default(),
-            signal_map: Default::default(),
-            default_reset_duration: 3,
-        }
-    }
-
-    fn build_event_map(&mut self, stmts: &[Statement]) {
-        let mut clock_insts: Vec<StrId> = Vec::new();
-        let mut reset_insts: Vec<StrId> = Vec::new();
-        Self::scan_tb_methods(stmts, &mut clock_insts, &mut reset_insts);
-        let program = self.program;
-        for inst in clock_insts.iter().chain(reset_insts.iter()) {
-            let name = veryl_parser::resource_table::get_str_value(*inst).unwrap_or_default();
-            if let Ok(addr) = program.get_addr(&[], &[&name]) {
-                self.event_map.insert(*inst, addr);
-                if let Some(info) = program.get_variable_info(&addr) {
-                    self.signal_map.insert(
-                        *inst,
-                        SemanticSignal {
-                            address: addr,
-                            width: info.width,
-                        },
-                    );
-                }
-            }
-        }
-    }
-
-    fn scan_tb_methods(stmts: &[Statement], clks: &mut Vec<StrId>, rsts: &mut Vec<StrId>) {
-        for stmt in stmts {
-            match stmt {
-                Statement::TbMethodCall(tb) => match &tb.method {
-                    TbMethod::ClockNext { .. } => {
-                        if !clks.contains(&tb.inst) {
-                            clks.push(tb.inst);
-                        }
-                    }
-                    TbMethod::ResetAssert { clock, .. } => {
-                        if !rsts.contains(&tb.inst) {
-                            rsts.push(tb.inst);
-                        }
-                        if !clks.contains(clock) {
-                            clks.push(*clock);
-                        }
-                    }
-                    TbMethod::FileOpen { .. }
-                    | TbMethod::FileWrite { .. }
-                    | TbMethod::FileClose
-                    | TbMethod::FileFlush => {}
-                },
-                Statement::If(s) => {
-                    Self::scan_tb_methods(&s.true_side, clks, rsts);
-                    Self::scan_tb_methods(&s.false_side, clks, rsts);
-                }
-                Statement::For(s) => Self::scan_tb_methods(&s.body, clks, rsts),
-                _ => {}
-            }
-        }
-    }
-
-    fn convert(&mut self, stmts: &[Statement]) -> Vec<SemanticStatement<AbsoluteAddr>> {
-        let p = self.program;
-        let root_instance_id = *p
-            .frontend
-            .instance_ids
-            .get(&crate::ir::InstancePath(Vec::new()))
-            .expect("root instance not found");
-        let root_module_id = p.frontend.instance_module[&root_instance_id];
-        let ec = ExprCompiler {
-            program: self.program,
-            testbench_source: self.testbench_source,
-            root_instance_id,
-            root_module_id,
-        };
-        let site_count = count_assert_statements(stmts, &self.testbench_source.functions) as u32;
-        let mut next_assert_site_id = p
-            .runtime_schema
-            .runtime_event_sites
-            .len()
-            .saturating_sub(site_count as usize) as u32;
-        stmts
-            .iter()
-            .filter_map(|s| self.convert_stmt(s, &ec, &mut next_assert_site_id))
-            .collect()
-    }
-
-    fn convert_stmt(
-        &self,
-        stmt: &Statement,
-        ec: &ExprCompiler<'_>,
-        next_assert_site_id: &mut u32,
-    ) -> Option<SemanticStatement<AbsoluteAddr>> {
-        fn convert_for_bound(
-            bound: &ForBound,
-            ec: &ExprCompiler<'_>,
-        ) -> GenericLoopBound<ExprBytecode<StateLocation<AbsoluteAddr>>> {
-            match bound {
-                ForBound::Const(x) => GenericLoopBound::Static(*x),
-                ForBound::Expression(expr) => GenericLoopBound::Dynamic {
-                    expr: ec.compile(expr.as_ref()),
-                    width: ec.root_context(expr).width,
-                    signed: expression_signed(expr),
-                },
-            }
-        }
-
-        match stmt {
-            Statement::TbMethodCall(tb) => self.convert_tb_method(tb, ec),
-            Statement::SystemFunctionCall(sf) => match &sf.kind {
-                SystemFunctionKind::Assert { kind, cond, args } => {
-                    let site_id = *next_assert_site_id;
-                    *next_assert_site_id = next_assert_site_id.saturating_add(1);
-                    Some(GenericTestbenchStatement::Assert {
-                        expr: ec.compile(&cond.0),
-                        site_id,
-                        continue_on_fail: matches!(kind, AssertKind::Continue),
-                        message: compile_assert_message(args, ec),
-                        location: extract_source_location(&sf.comptime.token),
-                    })
-                }
-                SystemFunctionKind::Finish => Some(GenericTestbenchStatement::Finish),
-                SystemFunctionKind::Display(args) => Some(GenericTestbenchStatement::Display {
-                    message: compile_assert_message(args, ec),
-                    newline: true,
-                }),
-                SystemFunctionKind::Write(args) => Some(GenericTestbenchStatement::Display {
-                    message: compile_assert_message(args, ec),
-                    newline: false,
-                }),
-                _ => None,
-            },
-            Statement::If(s) => Some(GenericTestbenchStatement::If {
-                expr: ec.compile(&s.cond),
-                then_block: s
-                    .true_side
-                    .iter()
-                    .filter_map(|s| self.convert_stmt(s, ec, next_assert_site_id))
-                    .collect(),
-                else_block: s
-                    .false_side
-                    .iter()
-                    .filter_map(|s| self.convert_stmt(s, ec, next_assert_site_id))
-                    .collect(),
-            }),
-            Statement::For(s) => {
-                let body: Vec<_> = s
-                    .body
-                    .iter()
-                    .filter_map(|s| self.convert_stmt(s, ec, next_assert_site_id))
-                    .collect();
-                let lv = self
-                    .resolve_loop_var(&s.var_id)
-                    .map(|(sig, width)| (sig, width, s.var_type.signed));
-                match &s.range {
-                    ForRange::Forward {
-                        start,
-                        end,
-                        inclusive,
-                        step,
-                    } => Some(GenericTestbenchStatement::For {
-                        loop_var: lv,
-                        start: convert_for_bound(start, ec),
-                        end: convert_for_bound(end, ec),
-                        inclusive: *inclusive,
-                        step: *step,
-                        step_op: None,
-                        reverse: false,
-                        body,
-                    }),
-                    ForRange::Reverse {
-                        start,
-                        end,
-                        inclusive,
-                        step,
-                    } => Some(GenericTestbenchStatement::For {
-                        loop_var: lv,
-                        start: convert_for_bound(start, ec),
-                        end: convert_for_bound(end, ec),
-                        inclusive: *inclusive,
-                        step: *step,
-                        step_op: None,
-                        reverse: true,
-                        body,
-                    }),
-                    ForRange::Stepped {
-                        start,
-                        end,
-                        inclusive,
-                        step,
-                        op,
-                    } => Some(GenericTestbenchStatement::For {
-                        loop_var: lv,
-                        start: convert_for_bound(start, ec),
-                        end: convert_for_bound(end, ec),
-                        inclusive: *inclusive,
-                        step: *step,
-                        step_op: Some(lower_testbench_operator(*op)),
-                        reverse: false,
-                        body,
-                    }),
-                }
-            }
-            Statement::Assign(a) => a
-                .dst
-                .first()
-                .and_then(|d| ec.resolve_var(&d.id))
-                .map(|dst| GenericTestbenchStatement::Assign {
-                    dst,
-                    expr: ec.compile_with_width(&a.expr, dst.width),
-                }),
-            Statement::Break => Some(GenericTestbenchStatement::Break),
-            Statement::FunctionCall(fc) => self.convert_function_call(fc, ec, next_assert_site_id),
-            _ => None,
-        }
-    }
-
-    /// Inline-expand a function call by binding arguments and converting
-    /// the function body's statements.
-    fn convert_function_call(
-        &self,
-        fc: &veryl_analyzer::ir::FunctionCall,
-        ec: &ExprCompiler<'_>,
-        next_assert_site_id: &mut u32,
-    ) -> Option<SemanticStatement<AbsoluteAddr>> {
-        let func = self.testbench_source.functions.get(&fc.id)?;
-        let func_body = if let Some(idx) = &fc.index {
-            func.get_function(idx)?
-        } else {
-            func.get_function(&[])?
-        };
-
-        // Build a list of statements: argument assignments + body
-        let mut stmts: Vec<SemanticStatement<AbsoluteAddr>> = Vec::new();
-
-        // Bind input arguments
-        for (arg_path, arg_expr) in &fc.inputs {
-            if let Some(&arg_var_id) = func_body.arg_map.get(arg_path) {
-                if let Some(sig) = ec.resolve_var(&arg_var_id) {
-                    stmts.push(GenericTestbenchStatement::Assign {
-                        dst: sig,
-                        expr: ec.compile_with_width(arg_expr, sig.width),
-                    });
-                }
-            }
-        }
-
-        // Inline body statements
-        for stmt in &func_body.statements {
-            if let Some(ts) = self.convert_stmt(stmt, ec, next_assert_site_id) {
-                stmts.push(ts);
-            }
-        }
-
-        if stmts.len() == 1 {
-            Some(stmts.into_iter().next().unwrap())
-        } else {
-            // Wrap multiple statements into an If(true) block as a sequence container
-            // (there's no "Block" variant in TestbenchStatement)
-            // Actually, we can return None and use a different approach:
-            // flatten into the parent's statement list.
-            // For now, wrap in an always-true If:
-            Some(GenericTestbenchStatement::If {
-                expr: ExprBytecode::new(vec![TbOpcode::ConstU64(1)]),
-                then_block: stmts,
-                else_block: Vec::new(),
-            })
-        }
-    }
-
-    fn convert_tb_method(
-        &self,
-        tb: &TbMethodCall,
-        ec: &ExprCompiler<'_>,
-    ) -> Option<SemanticStatement<AbsoluteAddr>> {
-        match &tb.method {
-            TbMethod::ClockNext { count, .. } => {
-                let ev = self.event_map.get(&tb.inst).copied()?;
-                let clock_count = match count {
-                    Some(expr) => {
-                        if let Some(n) = try_eval_const(expr) {
-                            GenericClockCount::Static(n)
-                        } else {
-                            GenericClockCount::Dynamic(ec.compile(expr))
-                        }
-                    }
-                    None => GenericClockCount::Static(1),
-                };
-                Some(GenericTestbenchStatement::ClockNext {
-                    clock_event: ev,
-                    count: clock_count,
-                })
-            }
-            TbMethod::ResetAssert { clock, duration } => {
-                let reset_signal = self.signal_map.get(&tb.inst).copied()?;
-                let clock_event = self.event_map.get(clock).copied()?;
-                let dur = duration
-                    .as_ref()
-                    .and_then(try_eval_const)
-                    .unwrap_or(self.default_reset_duration);
-                // Determine reset polarity from the variable's DomainKind
-                let (assert_value, deassert_value) = self.resolve_reset_polarity(&tb.inst);
-                Some(GenericTestbenchStatement::ResetAssert {
-                    reset_signal,
-                    clock_event,
-                    duration: dur,
-                    assert_value,
-                    deassert_value,
-                })
-            }
-            TbMethod::FileOpen { .. }
-            | TbMethod::FileWrite { .. }
-            | TbMethod::FileClose
-            | TbMethod::FileFlush => None,
-        }
-    }
-
-    /// Determine reset assert/deassert values from the variable's PortTypeKind.
-    /// PortTypeKind covers all four reset types (async/sync × high/low),
-    /// unlike DomainKind which maps sync resets to Other.
-    fn resolve_reset_polarity(&self, inst: &StrId) -> (u8, u8) {
-        let name = veryl_parser::resource_table::get_str_value(*inst).unwrap_or_default();
-        let program = self.program;
-        if let Ok(addr) = program.get_addr(&[], &[&name]) {
-            if let Some(info) = program.get_variable_info(&addr) {
-                return match info.type_kind {
-                    crate::ir::PortTypeKind::ResetAsyncHigh
-                    | crate::ir::PortTypeKind::ResetSyncHigh => (1, 0),
-                    crate::ir::PortTypeKind::ResetAsyncLow
-                    | crate::ir::PortTypeKind::ResetSyncLow => (0, 1),
-                    _ => (0, 1),
-                };
-            }
-        }
-        (0, 1)
-    }
-
-    fn resolve_loop_var(&self, var_id: &VarId) -> Option<(SemanticSignal<AbsoluteAddr>, usize)> {
-        let p = self.program;
-        let rid = p
-            .frontend
-            .instance_ids
-            .get(&crate::ir::InstancePath(Vec::new()))?;
-        let mid = p.frontend.instance_module.get(rid)?;
-        let vars = p.frontend.module_variables.get(mid)?;
-        let info = vars.get(var_id)?;
-        let addr = p.state_address_for_source(*rid, *var_id)?;
-        Some((
-            SemanticSignal {
-                address: addr,
-                width: info.width,
-            },
-            info.width,
-        ))
-    }
-}
-
-fn try_eval_const(expr: &Expression) -> Option<u64> {
-    match expr {
-        Expression::Term(f) => match f.as_ref() {
-            Factor::Value(c) => c.get_value().ok().map(|v| v.payload_u64()),
-            Factor::Variable(_, _, _, c) => c.get_value().ok().map(|v| v.payload_u64()),
-            _ => None,
-        },
-        _ => None,
-    }
-}
-
-// ── Helpers ────────────────────────────────────────────────────────────
-
-fn extract_source_location(
-    token: &veryl_parser::token_range::TokenRange,
-) -> Option<SourceLocation> {
-    let t = &token.beg;
-    let file = t
-        .source
-        .get_path()
-        .and_then(resource_table::get_path_value)?;
-    Some(SourceLocation {
-        file: file.to_string_lossy().into_owned(),
-        line: t.line,
-        column: t.column,
-    })
-}
 
 // ── Executor ───────────────────────────────────────────────────────────
 
@@ -3176,197 +719,11 @@ pub(crate) fn compile_semantic_testbench(
     program: &Program,
     source: &VerylTestbenchSource,
 ) -> Option<TestbenchProgram<AbsoluteAddr>> {
-    let initial_stmts = source.initial_statements.as_ref()?;
-    let mut builder = SemanticTestbenchBuilder::new(program, source);
-    builder.build_event_map(initial_stmts);
-    Some(TestbenchProgram::new(builder.convert(initial_stmts)))
-}
-
-fn bind_expr<B: SimBackend>(
-    sim: &Simulator<B>,
-    expr: ExprBytecode<StateLocation<AbsoluteAddr>>,
-) -> Option<CompiledExpr> {
-    let layout = sim.backend_ref().layout();
-    let bytecode = expr
-        .bind_with(|address| layout.offsets.get(address).copied())
-        .ok()?;
-    Some(CompiledExpr { bytecode })
-}
-
-fn bind_assert_arg<B: SimBackend>(
-    sim: &Simulator<B>,
-    arg: SemanticArgument<AbsoluteAddr>,
-) -> Option<CompiledAssertArg> {
-    Some(CompiledAssertArg {
-        expr: bind_expr(sim, arg.expr)?,
-        width: arg.width,
-        signed: arg.signed,
-        is_string: arg.is_string,
-    })
-}
-
-fn bind_assert_message<B: SimBackend>(
-    sim: &Simulator<B>,
-    message: GenericAssertMessage<SemanticArgument<AbsoluteAddr>>,
-) -> Option<AssertMessage> {
-    match message {
-        GenericAssertMessage::Formatted { template, args } => {
-            let args = args
-                .into_iter()
-                .map(|arg| bind_assert_arg(sim, arg))
-                .collect::<Option<Vec<_>>>()?;
-            Some(GenericAssertMessage::Formatted { template, args })
-        }
-        GenericAssertMessage::DynamicArgs(args) => {
-            let args = args
-                .into_iter()
-                .map(|arg| bind_assert_arg(sim, arg))
-                .collect::<Option<Vec<_>>>()?;
-            Some(GenericAssertMessage::DynamicArgs(args))
-        }
-    }
-}
-
-fn bind_clock_count<B: SimBackend>(
-    sim: &Simulator<B>,
-    count: GenericClockCount<ExprBytecode<StateLocation<AbsoluteAddr>>>,
-) -> Option<ClockCount> {
-    match count {
-        GenericClockCount::Static(count) => Some(GenericClockCount::Static(count)),
-        GenericClockCount::Dynamic(expr) => Some(GenericClockCount::Dynamic(bind_expr(sim, expr)?)),
-    }
-}
-
-fn bind_loop_bound<B: SimBackend>(
-    sim: &Simulator<B>,
-    bound: GenericLoopBound<ExprBytecode<StateLocation<AbsoluteAddr>>>,
-) -> Option<LoopBound> {
-    match bound {
-        GenericLoopBound::Static(bound) => Some(GenericLoopBound::Static(bound)),
-        GenericLoopBound::Dynamic {
-            expr,
-            width,
-            signed,
-        } => Some(GenericLoopBound::Dynamic {
-            expr: bind_expr(sim, expr)?,
-            width,
-            signed,
-        }),
-    }
-}
-
-fn bind_statement<B: SimBackend>(
-    sim: &Simulator<B>,
-    statement: SemanticStatement<AbsoluteAddr>,
-) -> Option<TestbenchStatement<B>> {
-    match statement {
-        GenericTestbenchStatement::ClockNext { clock_event, count } => {
-            Some(GenericTestbenchStatement::ClockNext {
-                clock_event: sim.backend_ref().resolve_event_opt(&clock_event)?,
-                count: bind_clock_count(sim, count)?,
-            })
-        }
-        GenericTestbenchStatement::ResetAssert {
-            reset_signal,
-            clock_event,
-            duration,
-            assert_value,
-            deassert_value,
-        } => Some(GenericTestbenchStatement::ResetAssert {
-            reset_signal: sim.backend_ref().resolve_signal(&reset_signal.address),
-            clock_event: sim.backend_ref().resolve_event_opt(&clock_event)?,
-            duration,
-            assert_value,
-            deassert_value,
-        }),
-        GenericTestbenchStatement::Assert {
-            expr,
-            site_id,
-            continue_on_fail,
-            message,
-            location,
-        } => Some(GenericTestbenchStatement::Assert {
-            expr: bind_expr(sim, expr)?,
-            site_id,
-            continue_on_fail,
-            message: match message {
-                Some(message) => Some(bind_assert_message(sim, message)?),
-                None => None,
-            },
-            location,
-        }),
-        GenericTestbenchStatement::Display { message, newline } => {
-            Some(GenericTestbenchStatement::Display {
-                message: match message {
-                    Some(message) => Some(bind_assert_message(sim, message)?),
-                    None => None,
-                },
-                newline,
-            })
-        }
-        GenericTestbenchStatement::If {
-            expr,
-            then_block,
-            else_block,
-        } => Some(GenericTestbenchStatement::If {
-            expr: bind_expr(sim, expr)?,
-            then_block: then_block
-                .into_iter()
-                .map(|statement| bind_statement(sim, statement))
-                .collect::<Option<Vec<_>>>()?,
-            else_block: else_block
-                .into_iter()
-                .map(|statement| bind_statement(sim, statement))
-                .collect::<Option<Vec<_>>>()?,
-        }),
-        GenericTestbenchStatement::For {
-            loop_var,
-            start,
-            end,
-            inclusive,
-            step,
-            step_op,
-            reverse,
-            body,
-        } => Some(GenericTestbenchStatement::For {
-            loop_var: loop_var.map(|(signal, width, signed)| {
-                (
-                    sim.backend_ref().resolve_signal(&signal.address),
-                    width,
-                    signed,
-                )
-            }),
-            start: bind_loop_bound(sim, start)?,
-            end: bind_loop_bound(sim, end)?,
-            inclusive,
-            step,
-            step_op,
-            reverse,
-            body: body
-                .into_iter()
-                .map(|statement| bind_statement(sim, statement))
-                .collect::<Option<Vec<_>>>()?,
-        }),
-        GenericTestbenchStatement::Assign { dst, expr } => {
-            Some(GenericTestbenchStatement::Assign {
-                dst: sim.backend_ref().resolve_signal(&dst.address),
-                expr: bind_expr(sim, expr)?,
-            })
-        }
-        GenericTestbenchStatement::Break => Some(GenericTestbenchStatement::Break),
-        GenericTestbenchStatement::Finish => Some(GenericTestbenchStatement::Finish),
-    }
-}
-
-fn bind_testbench_program<B: SimBackend>(
-    sim: &Simulator<B>,
-    program: TestbenchProgram<AbsoluteAddr>,
-) -> Option<Vec<TestbenchStatement<B>>> {
-    program
-        .into_statements()
-        .into_iter()
-        .map(|statement| bind_statement(sim, statement))
-        .collect()
+    celox_frontend_veryl::compile_semantic_testbench(
+        &program.frontend,
+        source,
+        program.runtime_schema.runtime_event_sites.len(),
+    )
 }
 
 /// Compile the root module's initial block into an executable native testbench.
@@ -3374,9 +731,7 @@ pub fn compile_initial_testbench<B: SimBackend>(
     sim: &Simulator<B>,
 ) -> Option<CompiledTestbench<B>> {
     let semantic = sim.program().testbench.clone()?;
-    Some(CompiledTestbench {
-        stmts: bind_testbench_program(sim, semantic)?,
-    })
+    celox_runtime::bind_testbench_program(sim.backend_ref(), semantic)
 }
 
 /// Execute a previously compiled native testbench against a built simulator.
@@ -3384,7 +739,7 @@ pub fn run_compiled_testbench<B: SimBackend>(
     sim: &mut Simulator<B>,
     tb: &CompiledTestbench<B>,
 ) -> TestResult {
-    run_testbench(sim, &tb.stmts)
+    run_testbench(sim, tb.statements())
 }
 
 /// Execute at most `tick_limit` simulator ticks from a compiled testbench.
@@ -3396,7 +751,7 @@ pub fn run_compiled_testbench_with_tick_limit<B: SimBackend>(
     tb: &CompiledTestbench<B>,
     tick_limit: u64,
 ) -> LimitedTestbenchResult {
-    run_testbench_limited(sim, &tb.stmts, Some(tick_limit))
+    run_testbench_limited(sim, tb.statements(), Some(tick_limit))
 }
 
 /// Run the testbench and return assertion results observed before the test
@@ -3865,15 +1220,15 @@ mod tests {
         let tb = compile_initial_testbench(&sim).unwrap();
 
         assert!(matches!(
-            tb.stmts.first(),
+            tb.statements().first(),
             Some(GenericTestbenchStatement::Display { newline: true, .. })
         ));
         assert!(matches!(
-            tb.stmts.get(1),
+            tb.statements().get(1),
             Some(GenericTestbenchStatement::Display { newline: false, .. })
         ));
         assert!(matches!(
-            tb.stmts.get(2),
+            tb.statements().get(2),
             Some(GenericTestbenchStatement::Finish)
         ));
     }
